Migrate persist docs to an adapter zone variant.
