+++ dev-requirements.txt
@@ -1,8 +1,8 @@
 # install latest changes in dbt-core + dbt-postgres
 # TODO: how to switch from HEAD to x.y.latest branches after minor releases?
-git+https://github.com/dbt-labs/dbt-core.git#egg=dbt-core&subdirectory=core
-git+https://github.com/dbt-labs/dbt-core.git#egg=dbt-tests-adapter&subdirectory=tests/adapter
-git+https://github.com/dbt-labs/dbt-core.git#egg=dbt-postgres&subdirectory=plugins/postgres
+git+https://github.com/dbt-labs/dbt-core.git@69846369c5d5d69eb50b565bcc684159bd7eefa4#egg=dbt-core&subdirectory=core
+git+https://github.com/dbt-labs/dbt-core.git@69846369c5d5d69eb50b565bcc684159bd7eefa4#egg=dbt-tests-adapter&subdirectory=tests/adapter
+git+https://github.com/dbt-labs/dbt-core.git@69846369c5d5d69eb50b565bcc684159bd7eefa4#egg=dbt-postgres&subdirectory=plugins/postgres
 
 black~=22.8.0
 click~=8.1.3
