--- tests/integration/persist_docs_tests/models-column-missing/missing_column.sql
@@ -1,2 +1,0 @@
-{{ config(materialized='table') }}
-select 1 as id, 'Ed' as name
