--- tests/integration/persist_docs_tests/models-column-missing/schema.yml
@@ -1,8 +1,0 @@
-version: 2
-models:
-  - name: missing_column
-    columns:
-      - name: id
-        description: "test id column description"
-      - name: column_that_does_not_exist
-        description: "comment that cannot be created"
