--- tests/integration/persist_docs_tests/models/my_fun_docs.md
@@ -1,10 +1,0 @@
-{% docs my_fun_doc %}
-name Column description "with double quotes"
-and with 'single  quotes' as welll as other;
-'''abc123'''
-reserved -- characters
---
-/* comment */
-Some $lbl$ labeled $lbl$ and $$ unlabeled $$ dollar-quoting
-
-{% enddocs %}
