--- tests/integration/persist_docs_tests/models/no_docs_model.sql
@@ -1,1 +1,0 @@
-select 1 as id, 'Alice' as name
