--- tests/integration/persist_docs_tests/models/schema.yml
@@ -1,70 +1,0 @@
-version: 2
-
-models:
-  - name: table_model
-    description: |
-      Table model description "with double quotes"
-      and with 'single  quotes' as welll as other;
-      '''abc123'''
-      reserved -- characters
-      --
-      /* comment */
-      Some $lbl$ labeled $lbl$ and $$ unlabeled $$ dollar-quoting
-    columns:
-      - name: id
-        description: |
-          id Column description "with double quotes"
-          and with 'single  quotes' as welll as other;
-          '''abc123'''
-          reserved -- characters
-          --
-          /* comment */
-          Some $lbl$ labeled $lbl$ and $$ unlabeled $$ dollar-quoting
-      - name: name
-        description: |
-          Some stuff here and then a call to
-          {{ doc('my_fun_doc')}}
-  - name: view_model
-    description: |
-      View model description "with double quotes"
-      and with 'single  quotes' as welll as other;
-      '''abc123'''
-      reserved -- characters
-      --
-      /* comment */
-      Some $lbl$ labeled $lbl$ and $$ unlabeled $$ dollar-quoting
-    columns:
-      - name: id
-        description: |
-          id Column description "with double quotes"
-          and with 'single  quotes' as welll as other;
-          '''abc123'''
-          reserved -- characters
-          --
-          /* comment */
-          Some $lbl$ labeled $lbl$ and $$ unlabeled $$ dollar-quoting
-
-seeds:
-  - name: seed
-    description: |
-      Seed model description "with double quotes"
-      and with 'single  quotes' as welll as other;
-      '''abc123'''
-      reserved -- characters
-      --
-      /* comment */
-      Some $lbl$ labeled $lbl$ and $$ unlabeled $$ dollar-quoting
-    columns:
-      - name: id
-        description: |
-          id Column description "with double quotes"
-          and with 'single  quotes' as welll as other;
-          '''abc123'''
-          reserved -- characters
-          --
-          /* comment */
-          Some $lbl$ labeled $lbl$ and $$ unlabeled $$ dollar-quoting
-      - name: name
-        description: |
-          Some stuff here and then a call to
-          {{ doc('my_fun_doc')}}
