--- tests/integration/persist_docs_tests/models/table_model.sql
@@ -1,2 +1,0 @@
-{{ config(materialized='table') }}
-select 1 as id, 'Joe' as name
