--- tests/integration/persist_docs_tests/models/view_model.sql
@@ -1,2 +1,0 @@
-{{ config(materialized='view') }}
-select 2 as id, 'Bob' as name
