--- tests/integration/persist_docs_tests/seeds/seed.csv
@@ -1,3 +1,0 @@
-id,name
-1,Alice
-2,Bob
